added restriction to the admin page

+++ src/components/adminPage/teamDetails.js
@@ -7,33 +7,39 @@ import { useNavigate } from "react-router-dom";
 const TeamDetails =({users, teams})=>{
   const {alert, showAlert} = useAlert();
   const navigate = useNavigate();
+  const token = localStorage.getItem("adminAuthToken");
+  const team = localStorage.getItem("team")
 
   const handleDelete = async(id)=>{
-    const token = localStorage.getItem("adminAuthToken");
+    if(team.role === "Admin"){
+      try {
+        const response = await fetch(`http://localhost:4000/admin/team-delete/${id}`, {
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
 
-    try {
-      const response = await fetch(`http://localhost:4000/admin/team-delete/${id}`, {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
+        const result = await response.json();
+        if(response.ok){
+          showAlert(result.message || "Team deleted succesfully", "success");
+          navigate("/admin/team")
+        }else{
+          showAlert("Failed to delete selected team", "error")
+        }
+        
+      } catch (error) {
+        console.error("Error deleting team:", error)
+        if(error.response && error.response.status === 500){
+          navigate("/server-error")
+        }else{
+          showAlert("Error deleting team member ", "error");
+        }
+      }
 
-      const result = await response.json();
-      if(response.ok){
-        showAlert(result.message || "Team deleted succesfully", "success");
-        navigate("/admin/team")
-      }else{
-        showAlert("Failed to delete selected team", "error")
-      }
-      
-    } catch (error) {
-      console.error("Error deleting team:", error)
-      if(error.response && error.response.status === 500){
-        navigate("/server-error")
-      }else{
-        showAlert("Error deleting team member ", "error");
-      }
+    }else{
+      return;
+
     }
   }
 
@@ -77,7 +83,7 @@ const TeamDetails =({users, teams})=>{
                 <h6 className="clockout_column">{matchingUser.email_address}</h6>
                 <h6 className="hours_column">{team.staff_code}</h6>
                 <h6 className="status_column">
-                  <Link to={`/admin/team/${team._id}`}><img src="/icons/Edit.svg" alt="" /></Link>
+                  <Link to={ team.role === "Admin" ? `/admin/team/${team._id}` : ""}><img src="/icons/Edit.svg" alt="" /></Link>
                   <img
                     src="/icons/Delete.svg" 
                     alt="Delete-icon"
